Using a dictionnary to store the values of each pixel
by storing the grayscale values of the pixels, I will be able to create a thresholding that works for any picture. The pixels are sorted by grayscale value, which means that I can choose a threshold value based on a certain percentage of pixels that are under that value. This guarantees that the colors are separated somewhat equally in every picture, preventing saturation. This might cause problems if an image has a large quantity of big speckles, but I plan on having a way to easily adjust the percentages in cas there is a problem.