adjusted thresholding
finished the adjusted thresholding. I will find a way to make it easier to change the parameters. I also added a way to get a complete thresholding, which will help for images that aren't in focus, since there is a lot of values that we do not want while doing the hough transform.